test: move metadata serialization test to jvm
