+++ bugsnag-android-core/src/test/resources/meta_data_serialization_2.json
@@ -1,0 +1,9 @@
+{
+  "example": {
+    "map": {
+      "key": {
+        "key": "value"
+      }
+    }
+  }
+}
