+++ bugsnag-android-core/src/test/resources/meta_data_serialization_3.json
@@ -1,0 +1,10 @@
+{
+  "example": {
+    "list": [
+      [
+        "james",
+        "test"
+      ]
+    ]
+  }
+}
