+++ bugsnag-android-core/src/test/resources/meta_data_serialization_4.json
